Add --no-update option
A singular release is a komodo release which isn't tracked by any of:
stable, testing, unstable or bleeding. The bug was that
`resolve_release` failed if it couldn't find a tracking symlink
regardless of the user's intention.

For komodo integration tests, we create komodoenvs towards these
singular releases, since the unstable symlink isn't created before the
integration tests begin, v0.3 of komodoenv would fail. This commit adds
the `--no-update` option which disables tracking.

This is done by having `tracked-release == current-release`. When
`komodoenv-update` is called, it'll work transparently and be a no-op.